+++ src/com/ipi/glcm/NeighborCounter.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public final class NeighborCounter {
 
-    public static void countOccurrences(List<List<Integer>> image, GlcmFeatures matrix) {
+    public static void countOccurrences(List<List<Double>> image, GlcmFeatures matrix) {
         for (int i = 0; i < image.size(); i++) {
             for (int j = 0; j < image.get(0).size(); j++) {
                 List<Integer> current = makeList(i, j);
@@ -17,7 +17,7 @@ public final class NeighborCounter {
         }
     }
 
-    private static void countNeighborhood(List<List<Integer>> image, GlcmFeatures matrix, List<Integer> current) {
+    private static void countNeighborhood(List<List<Double>> image, GlcmFeatures matrix, List<Integer> current) {
         zeroDegreeCounter(image, current, matrix.getZeroDegree());
         fortyFiveDegreeCounter(image, current, matrix.getFortyFiveDegree());
         ninetyDegreeCounter(image, current, matrix.getNinetyDegree());
@@ -25,37 +25,37 @@ public final class NeighborCounter {
     }
 
 
-    private static void zeroDegreeCounter(List<List<Integer>> image, List<Integer> current, GlcmMatrix glcmMatrix) {
+    private static void zeroDegreeCounter(List<List<Double>> image, List<Integer> current, GlcmMatrix glcmMatrix) {
         if ((current.get(1) + glcmMatrix.getD()) < image.get(0).size()) {
-            int i = image.get(current.get(0)).get(current.get(1));
-            int j = image.get(current.get(0)).get(current.get(1) + glcmMatrix.getD());
+            int i = image.get(current.get(0)).get(current.get(1)).intValue();
+            int j = image.get(current.get(0)).get(current.get(1) + glcmMatrix.getD()).intValue();
             glcmMatrix.incrementElement(i, j);
         }
     }
 
-    private static void fortyFiveDegreeCounter(List<List<Integer>> image, List<Integer> current,
+    private static void fortyFiveDegreeCounter(List<List<Double>> image, List<Integer> current,
             GlcmMatrix glcmMatrix) {
         if (current.get(0) - glcmMatrix.getD() >=  0 && image.get(0).size()  - current.get(1) > glcmMatrix.getD()) {
-            int i = image.get(current.get(0)).get(current.get(1));
-            int j = image.get(current.get(0) - glcmMatrix.getD()).get(current.get(1) + glcmMatrix.getD());
+            int i = image.get(current.get(0)).get(current.get(1)).intValue();
+            int j = image.get(current.get(0) - glcmMatrix.getD()).get(current.get(1) + glcmMatrix.getD()).intValue();
             glcmMatrix.incrementElement(i, j);
         }
     }
 
-    private static void ninetyDegreeCounter(List<List<Integer>> image, List<Integer> current,
+    private static void ninetyDegreeCounter(List<List<Double>> image, List<Integer> current,
             GlcmMatrix glcmMatrix) {
         if (current.get(0) + glcmMatrix.getD() < image.size()) {
-            int i = image.get(current.get(0)).get(current.get(1));
-            int j = image.get(current.get(0) + glcmMatrix.getD()).get(current.get(1));
+            int i = image.get(current.get(0)).get(current.get(1)).intValue();
+            int j = image.get(current.get(0) + glcmMatrix.getD()).get(current.get(1)).intValue();
             glcmMatrix.incrementElement(i, j);
         }
     }
 
-    private static void oneHundredThirtyFiveCounter(List<List<Integer>> image, List<Integer> current,
+    private static void oneHundredThirtyFiveCounter(List<List<Double>> image, List<Integer> current,
             GlcmMatrix glcmMatrix) {
         if (current.get(0) + glcmMatrix.getD() < image.size() && current.get(1) + glcmMatrix.getD() < image.get(0).size()) {
-            int i = image.get(current.get(0)).get(current.get(1));
-            int j = image.get(current.get(0) + glcmMatrix.getD()).get(current.get(1) + glcmMatrix.getD());
+            int i = image.get(current.get(0)).get(current.get(1)).intValue();
+            int j = image.get(current.get(0) + glcmMatrix.getD()).get(current.get(1) + glcmMatrix.getD()).intValue();
             glcmMatrix.incrementElement(i, j);
         }
     }
